PLANET-7849: Refactor header.js file
- Refactor header.js file
